+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## NEXT
 
+**Features**
+
+* Add support for `eth_call` for Senders in order to simulate sending the transaction and receive a result or parsed revert error (`@colony/colony-js-client`, `@colony/colony-js-contract-client`, `@colony/colony-js-adapter`, `@colony/colony-js-adapter-ethers`)
+* Add `reason` property to transaction receipts (`@colony/colony-js-contract-client`)
+
 **Bug fixes**
 
 * Add custom Infura provider which retries queries when bad responses are received, which happens occasionally (`@colony/colony-js-client`)